+++ tests/macros.rs
@@ -10,9 +10,9 @@ macro_rules! all_log_macros {
     });
 }
 
-struct NopLogger;
+struct Logger;
 
-impl Log for NopLogger {
+impl Log for Logger {
     fn enabled(&self, _: &Metadata) -> bool {
         false
     }
@@ -39,10 +39,10 @@ fn no_args() {
     all_log_macros!(target: "my_target", "hello");
     all_log_macros!(target: "my_target", "hello",);
 
-    all_log_macros!(logger: NopLogger, "hello");
-    all_log_macros!(logger: NopLogger, "hello",);
-    all_log_macros!(logger: NopLogger, target: "my_target", "hello");
-    all_log_macros!(logger: NopLogger, target: "my_target", "hello",);
+    all_log_macros!(logger: Logger, "hello");
+    all_log_macros!(logger: Logger, "hello",);
+    all_log_macros!(logger: Logger, target: "my_target", "hello");
+    all_log_macros!(logger: Logger, target: "my_target", "hello",);
 }
 
 #[test]
@@ -130,8 +130,8 @@ fn kv_no_args() {
         log!(lvl, cat_1 = "chashu", cat_2 = "nori", cat_count = 2; "hello");
     }
 
-    all_log_macros!(logger: NopLogger, cat_1 = "chashu", cat_2 = "nori", cat_count = 2; "hello");
-    all_log_macros!(logger: NopLogger, target: "my_target", cat_1 = "chashu", cat_2 = "nori", cat_count = 2; "hello");
+    all_log_macros!(logger: Logger, cat_1 = "chashu", cat_2 = "nori", cat_count = 2; "hello");
+    all_log_macros!(logger: Logger, target: "my_target", cat_1 = "chashu", cat_2 = "nori", cat_count = 2; "hello");
     all_log_macros!(target: "my_target", cat_1 = "chashu", cat_2 = "nori", cat_count = 2; "hello");
     all_log_macros!(target = "my_target", cat_1 = "chashu", cat_2 = "nori", cat_count = 2; "hello");
     all_log_macros!(cat_1 = "chashu", cat_2 = "nori", cat_count = 2; "hello");
